[REFACTOR] 디렉토리 통합 - 어르신 정보 코드 CommonController 로 통합
